+++ boltprojection/adaptor_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/dogmatiq/dogma"
 	. "github.com/dogmatiq/enginekit/enginetest/stubs"
 	. "github.com/dogmatiq/projectionkit/boltprojection"
-	"github.com/dogmatiq/projectionkit/boltprojection/fixtures" // can't dot-import due to conflict
+	"github.com/dogmatiq/projectionkit/boltprojection/internal/fixtures" // can't dot-import due to conflict
 	"github.com/dogmatiq/projectionkit/internal/adaptortest"
 	"github.com/dogmatiq/projectionkit/internal/identity"
 	. "github.com/onsi/ginkgo"